requirements.txt
eth-ape
silverback
slack_sdk
setuptools
ape-base
ape-alchemy

